requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.7
argon2-cffi==23.1.0
PyJWT==2.8.0
alembic==1.12.1
Flask-Migrate==4.0.5